+++ CHANGES.md
@@ -1,4 +1,8 @@
 ## [unreleased]
+### Added
+- make_pkg_name public api to convert an addon name to a python 
+  package name
+
 ### Fixed
 - crash in case a previous commit had a bad `__openerp__.py`